# Use binary search to further speed up