--- a/types/field_type.go
+++ b/types/field_type.go
@@ -162,18 +162,18 @@ func DefaultTypeForValue(value interface{}, tp *FieldType) {
 		SetBinChsClnFlag(tp)
 	case int:
 		tp.Tp = mysql.TypeLonglong
-		tp.Flen = len(strconv.FormatInt(int64(x), 10))
+		tp.Flen = LenOfInt64Fast(int64(x))
 		tp.Decimal = 0
 		SetBinChsClnFlag(tp)
 	case int64:
 		tp.Tp = mysql.TypeLonglong
-		tp.Flen = len(strconv.FormatInt(x, 10))
+		tp.Flen = LenOfInt64Fast(x)
 		tp.Decimal = 0
 		SetBinChsClnFlag(tp)
 	case uint64:
 		tp.Tp = mysql.TypeLonglong
 		tp.Flag |= mysql.UnsignedFlag
-		tp.Flen = len(strconv.FormatUint(x, 10))
+		tp.Flen = LenOfUint64Fast(x)
 		tp.Decimal = 0
 		SetBinChsClnFlag(tp)
 	case string: