Signed-off-by: Fuyang Liu <liufuyang@users.noreply.github.com>

To speed up this https://github.com/pingcap/tidb/pull/9726 even more

* Without change, as the version on pull 9726
```
BenchmarkDefaultTypeForValue/LenOfUint64_input_full_range-8         	50000000	        33.6 ns/op
BenchmarkDefaultTypeForValue/LenOfUint64_input_0_to_64K__-8         	100000000	        21.8 ns/op
BenchmarkDefaultTypeForValue/LenOfUint64_input_0_to_512__-8         	100000000	        22.2 ns/op
```

* With this PR, by @TennyZhuang's binary idea
```
BenchmarkDefaultTypeForValue/LenOfUint64_input_full_range-8         	50000000	        26.7 ns/op
BenchmarkDefaultTypeForValue/LenOfUint64_input_0_to_64K__-8         	100000000	        21.1 ns/op
BenchmarkDefaultTypeForValue/LenOfUint64_input_0_to_512__-8         	100000000	        21.2 ns/op
```

We could see some further speed improvements.

@ngaut can you (or let someone) review this to see if we should use this binary version as @TennyZhuang suggested? It is slightly faster.

More testing code and results can be seen here:
https://gist.github.com/liufuyang/a36de53033d989320664e5c242d7d4aa